Array Asserts: Converted inlace array asserts to print/stop() pattern
Previously the array functions asserts quit esh completely and printed out a line number (which changed often).
Asserts should not be used to make assertions about user programs. Stop() is a cleaner exit.
Messages still should be routed to the error queue to the context.  However, queues don't exist yet.
so printf() is used for now.