ArrayCore: ResizeCore is now ResizeCapacity, Simplified array construction, Array has Rank() Method.
Array constructor now simply resize to empty. this triggers any necessary allocation for bounded or fixed arrays.
The Arrays Rank Method still get the rank from the type, but callers no longer need to see that detail and makes it easier to
store the Rank in the array later if it helps.