TempStackCString : Refactored class so core array can be used for more cases.
Class used to be hardcoded to 255 Utf8Char elements + 1 for 0 termination. Now a template is used
internally so other fixed arrays can use this design patter as well.  This is a staging to allow a fixed sized
initial array that can spill to dynamic allocation if necessary. At that point, perhaps even the fixed size
initial buffer is not that helpful. Either way, trying to move more code to the common API pattern.