FileIO, UnitTests: Println and Printf both use the posix write function
previously on used printf under the hood and the otters used write. This resulted in different
buffering/ flushing behaviors for the two. Test results have been updated now that output ordering is consistent
Historically DPrintf was treated as a special case since it was "just" for debugging. That s no longer the case.