Contexts: TADMs now refer to ExecutionContexts so many TADS can share one exec.
This change in perspective makes its easier for the execution system to run slices from
clumps that come from many different contexts. It will still be possible to have separate
exec contexts for even more isolation. This is the second and a few steps to make it easier
to run multiple contexts efficiently in a JS runtime.

At this point only the top user context is cleared out. sub contexts still need to be cleaned up .