TDCodecVIA: More work on literal/data parsing
The distinction between data and literal is slim and getting smaller.
Literals are potential completely context free and are sued to generate a new symbol
(e.g. a TypeRef)  Data is directly parsed based on an and existing TypeRef the latter is
needed for clusters to that each field is parsed according to type it supposed to fill
the structure of an existing symbol.