TDCodecVia: Removed all '.' prefixes on structure values in vireo, they still remain in test programs.
The '.' has been optional for a while, so this is more of a cleanup task. The semantics of type creation
changed a while back to be based on using values as prototypes, instead of a value's type. That
change reduced the situations where the internal type concept surfaced in the via language.