Add Vireo function for InRangeAndCoerce.  Works on arrays, clusters, or scalars:
  InRangeAndCoerce(xValue loValue hiValue includeLo includeHi coercedXOut boolOut)

Computes whether xValue is inbetween loValue and hiValue (inclusive or exclusive
depending on Boolean flags for each).
boolOut output is either a single Boolean (Compare Aggregates mode, in which case
coercecXOut is the same as xValue and entire clusters or arrays are treated
as a single aggregate object), or a cluster or array of Booleans with the same
data shape as the inputs (e.g. array of clusters of 2 ints inputs = array of
clusters of 2 bools output).
Add InRangeAndCoerce.via test and results .vtr.