Initial support for enums
Example:
    e(dv(Enum16 ("abc def" %47hi jkl mno) 2) bar)
    e(dv(Enum32 (red orange yellow green blue purple)) foo)

Definition, string output and formatting with %d and %s work
Copy and Convert from int types to enums pin output to be in range.

Arithmetic prims currently do not pin, but having DFIR generate a
Copy from the enum to itself after arith ops would correct this in
the meantime.