withCredentials to false to enable Anon. Requests
Anonymous Requests require withCredentials to be false and is currently our most frequent use case in testing.

However, this will prevent making HTTP requests to Cross-Origin servers using Cookies or the HTTP Auth Header for authentication. So we need a way to expose configuration of this property to users.

It should not have an effect making requests to same origin servers.