Add CallChain prim.
Usgage:  CallChain(o(Array(String *)))

Returns an array of caller VIs, element zero being the most recent on
the 'stack'.  It supports reentrant VIs but does not yet support adding
":#" to the end, because we don't yet support clone pools.

Added a SubString field to VirtualInstrument so it can know its own
name.  (Previously this was only known to the type system via 'define').