Add Error I/O to StringFormat().
Allow named ErrorClust type to be used in prim descriptor strings
by enhancing argument/formal argument type matching code.

Allow multiple var arg prims with different initial types (before
the varargs portion) to resolve to the right prim based on these types
correctly, and use this to allow versions of StringFormat both with
and without error I/O to co-exist.  (In the future we may remove the
one without error I/O, but for now, this allows existing tests and DFIR
codegen to continue to function.)