Fix native test failures on Windows for tests with embedded \r in output.
Currently the only test this affects is StringTrim.via
Windows Console I/O outputs \r as a CR but also always
converts \n to CR LF.   So outputing the string \r\n actually output CR
CR LF.  This translation is done by the Windows Console I/O layer, not
by Vireo; using printf or fwrite does it.  It's not clear we want to go
setting stdout to binary mode or something, so for now this seems the
best way to deal with it.