Fix lint errors for TODO usage, avoiding using namespace,  and header/include-what-you-use
We now have zero lint errors except for the a couple of disabled ones.

The only lint errors still disabled are: enforcing open braces at end of
line, the format of the #ifdef include guard name, and C-style casts.
We also allow line to be up to 160 chars rather than 80.
See the comment at the top of make-it/cpplint/Makefile.