Prevent emsdk update (#299)
Running the emsdk update command causes the emsdk tool itself to update while used to install a specific version of emscripten.

This was unexpected and causes a latest incoming change to emsdk to (1) error during emsdk install resulting in breaking builds and (2) creates extremely verbose output making it difficult to look at build results.

An issue was filed here: https://github.com/juj/emsdk/issues/86

This change lets us sync to a specific version of the emsdk tool prior to those issues being introduced.

As the build is currently broken, this is a high priority change and as soon as the test builds go in I will merge this change. Please add review comments and they will be addressed before merge if possible or immediately after.