Fix Issue #310
ArrayInsertSubset was incorrectly limiting the dim length of
the inserted subarray to be no more than the length of the input
array along the inserted dimension.  All of the other dimensions
are correctly so-pinned (so the array grows along only one dimension), but
the inserted dimension should not be limited.
E.g.  Insert (by rows) into [[1,1][2,2]] a subarray of [3,3,3],[4,4,4],[5,5,5]]
The columns are pinned to 2 elements because the subarray has two many
columns to fit into the input array, but all 3 rows of subarray should
still be inserted into the input array, even though it has only 2 rows
to begin with, because the array grows along that dimension, but not
the others.