Provide a type to Blob so empty string is not used
Coherent uses the default empty string for the Blob type as the
Content-Type of the XHR which is out of spec:
https://fetch.spec.whatwg.org/#concept-bodyinit-extract

This causes Coherent to send XHR requests that have the Content-Type of
empty string. Here the default was changed to text/plain because it is
the most general of the three allowed headers for a Simple CORS Request:
https://www.w3.org/TR/cors/#simple-header

The charset was set to utf-8 as that is the default charset of Vireo
strings.

This seems like a reasonable default and tests should be added to
confirm the default is overridden if a new Content-Type is provided.