Fix a conversion issue between floats and unsigned ints.
The runtime engine in LabVIEW when asked to convert a negative
double/float into an unsigned integer of any size. To match this
behavior, I added a check in the Math.cpp file to coerce the source
value to the minimum representable value by the destination data type if
necessary.