Add default Content-Type if user does not provide
IE and Edge do not propagate the Content-Type from the Blob type so
instead check all of the headers added by the user and if they do not
add a Content-Type then add the default ourselves.

Unfortunately setRequestHeader is poorly named as it doesn't replace an
existing header but instead appends if a header of the same name was
already added. Because of this went with the flag approach instead.