Fix bugs in Spreadsheet String to Array
1) Fix a bug in SubString::FindFirstMatch which failed to find an exact
match at the tail end of a string (e.g. match 'u' in 'aeuou').
2) Fix matching of %s in Spreadsheet String to Array to not stop at
whitespace, since this prim has different meaning than Scan From String.
3) Add support for passing an array of delimiter strings
3) Intentionally replicate a minor bug in CG and desktop NCG that causes a
partial match of a delimiter string at the end of the input to ignore the
final delimiter when a complete match would occur later in the delim array.
(See final two cases in TestVI_MultipleDelimiters.via).
We should consider fixing this in both, but there's a danger of
breaking existing customer code in CG so it may be better to leave
this unaddressed and document it; parsing behavior would change
subtlely and return an extra default default element at the end of the
each row in this corner case.

Add tests.