CAR 666144 addendum for array of queues.
There's two issue here:

Issue 1:
CopyData wasn't correctly implemented for RefNumValType.  It was
incorrectly called the base class wrapped type, which treated the refnum
data as an array (which it's not) and crashing.  (The refnum does hold
onto an array type to manage the refnum's queue, but the array is stored
out of line in the refnum manager, accessed via refnum cookie, and
not directly in the data).  This part of the fix is straight-forward,
was obviously wrong before, and is safe.

Issue 2:
ArrayConcatenate emits an array of booleans after the input element arguments
to identify whether each of inputs is an array or a scalar element.
EmitArrayConcatenateInstruction() calls CompareType with the output type
for each input type to decide this.  CompareType(A,B) ultimately will
return true if either IsA(A,B) or IsA(B,A), and since a refnum's type is
built out of an array type so it can remember the element type and
possible static size of the queue, IsA(refnum, array) winds up returning
true.  The fix for this is to shortcircuit the base-type loop in IsA and
make it stop when it encounters a refnum, considering it opaque.
(Probably PointerType should do this too.)

(The IsA change explicitly tests for refnums, so it should be limited in
scope.  But it's a bit more risky than the other array of non-flat data fix
because it touches a very commonly used function.)