Disable the ill-conceived and senseless no-negated-condition eslint rule.
It's a semantic rule, not a style rule, and could actively hurt
code readability by forcing the inversion of if/else statements
into a less logical order to silence the warning.  Lint programs based
on simple pattern-matching, which have no actual understanding of the
code they are processing, have no business enforcing this type of rule.
There are many, many cases where a non-zero return value from a call
is the common, non-error condition, and you want the code which handles
this case first and primary.  There is a reason languages with
exception handling have try/catch blocks and not catch/try blocks.