Add setXMLHttpRequestImplementation and refactor abortAllRunningRequests
The abortAllRunningRequests function was exposed to the user as a per
instance Vireo method but effected all running Vireo instances.
Refactored abortAllRunningRequests to behave per instance.

Also added setXMLHttpRequestImplementation which can be used to swap the
underlying XMLHttpRequest implementation. This is valuable for the
node.js environment to avoid the need to set a global XHR implementation
and also valuable for testing purposes.