Expose ExecuteSlices millisecondsToRun
Changes ExecuteSlices so that instead of taking ticks to run and
returning ms to wait it takes ms to run and returns ms to wait. Modified
the js executeSlicesUntilWait entry point to passthrough milliseconds to
run.

Does not represent an api change as undefined parameters in this context
are coerced to zero (due to asm.js / wasm numeric coercison) and a value
of zero ms to wait is effectively the same as the previous value of 20
ticks to wait.