Fix to handle data values for optional arguments.
The existing code in ClumpParseState::ResolveActualArgument created a null
pointer for optional arguments for flat types but created a new value for
non-flat types. This was incorrect because the value for this optional
argument (though not explicit in the VIA code) is reused for subsequent
calls. (Almost) All the primitives explicitly handle correctly whether an
optional parameter is null by checking the pointer and using the correct
default value if null regardless of whether the type of the argument is
flat or not. Thus this method is fixed to treat flat and non-flat types
same and not allocate data value for optional arguments.

Add a test (WildcardArguments.via)

Some primitives such as StringFormat, FormatDateTimeString are the
exceptions that did not handle the optional arguments and they are fixed
to do so.

A memory overrun in StringScan is fixed by using AliasAssignCStrLen.

Fix InlineConstants.via to not use first argument as optional argument any
more as it is not anymore.