CAR 682384: Fix to handling parse errors within ParseArrayData
Fix for CAR 682384: WebVI hang executing Unflatten from JSON

ParseArrayData doesn't break out of its own parse loop when there is an error parsing elements. This can result in an infinite loop within the array's parse loop. If the parse of the element has an error, then we'll now break the loop and bubble the error up so it is returned by the ParseArrayData call.