Fixing additional ArrayDelete issues related to 1D arrays
I found some more ArrayDelete issues with 1D arrays.

First, offset (the starting index of the delete) should default to the length of the array minus the length of the delete as it does in Desktop. This means we always delete the last elements of the array without an offset specified.

Second, if the offset was larger than the array, cap it at the length, so that later on the startIndex and endIndex will be equal and we won't delete anything, which makes sense since the offset suggests we're trying to delete outside of the array. Not capping it like this could result in an attempt to access memory outside of the array, which could result in an access violation.

Finally, if a negative length was given, then endIndex may be less than startIndex which can result in the deletedArray having a size of -1. Any attempt to copy into that would result in an access violation.