Fix memory leak due to passing wildcard arguments in SubVI calls
Disallow writing to inputs in subVIs, which (because inputs are either
aliases to caller's data or allocated in parm block for wildcard args)
can overwrite the only reference to a wild/default'd allocation and
cause a leak.

When passing multiple args to the native esh command line program, run
them in the same TypeManager shell to stress test and help find leaks.

Turn on VIREO_ASSERT for native builds, even in release target, to better find
issues on the CI builds.