Implement control references.
Note: Control references are defined in the Locals: block of a VI (or inside an EventSpec),
rather than in a "ControlRefs:" block as we had discussed. (The reason is
that each ControlReference definition defines a local name which must be
visible as a local from within the clumps of the VI.)

Example:

define(MyVI dv(.VirtualInstrument (
    Locals: c(
        e(ControlReference("dataItem_Foo") ctlref1)  // static control ref linked to dataItem_Foo
    ) ...

The actual Type of ctlref1 is ControlRefNum and stores a refnum (cookie) which
opaquely holds onto the VI and control tag (data item).
The VI isn't passed directly, it's implicitly taken from the context,
whatever VI the ControlReference is defined in.  (Note, this is a refinement of
the original model we worked out which passed the VI's name as a string.)
ControlRefNums defined with ControlReference() exist during the entire
lifetime of the VI and are only deallocated when the last clump of the last
top-level VI finishes.

Non-statically-bound ControlRefNums variables can also be declared in the Params:
or Locals: sections, e.g.  e(ControlRefNum ref).
These can be used in subVIs to refer to a reference passed in by a caller.

See test-it/ViaTests/ControlRefBasic.via for examples.