Fix long-standing race-condition/corruption bug in Observables
Basically, if a prim which owns an Observable (such as Queue Enq/Deq,
Occurrence, or WaitOnEvent) waits on the observable and reshedules
itself, then another clump runs which calls ObserveStateChange to
reschedule the clump (for example, a WaitOnOccurrence in the first clump
and SetOccurrence in the second), BUT before the original clump can re-run,
the ExecSlices inner loop breaks out and runs CheckTimers, and by
coincidence the second clump ran long enough that the original prim
timed out, the timeout proc would cause the same clump to be scheduled AGAIN.

Our execution systems queues are not expecting a clump to ever be
scheduled more than once, and this corrupts the data structures and
could cause an assert, an infinite run loop, or an early termination
('losing' a clump).

This bug would manifest most often in an intermittent failure in QueueRef
tests on the CI, but I recently ran into it with an event test that was
more predictable and allowed me to track it down.

Fix is to make ObserveStateChange cancel its associated timer.
Also added assert to VIClump Enqueue to verify at least that the
inserted clump is not already the head or tail.