Add support for async calls through the JavaScriptInvoke function (#422)
Updated the JavaScriptInvoke method call to include an occurrence pointer, similar to the existing HTTP methods. With this change, synchronous calls to JavaScriptInvoke will also create an occurrence and wait for the occurrence to be set before completing the call. There was no significant performance impact on synchronous calls due to this PR.