Keep function context and add jsAPI parameter
Preserves the context when a function is retrieved from the global scope to use as the context for invocation. Exposes the JS API to the function if the user adds an additional parameter to the JS function call.

For instance, if the JSLI definition specifies four parameters but the JS function call specifies 5 parameters, the fifth parameter will be the JS API context. Two benefits are that we only need to define the context if the user requests it (by adding an addditional parameter) and this pattern potentially leaves us open for method invocation in the future where the user may provide a context.