Refactor RefNumVal type to not carry its type with its data.
Uses refactored to get the type from other parameters or a StaticType
argument.  This allows the EventSpec to stay the same size when
specified with a control ref or with a jsRef or control ID.

Added a new kind of StaticType argument, "StaticTypeExplicitData",
which is like StaticTypeAndData except you also specify the data
argument type separately too. This allows you to still have type checking
(e.g. a prim is polymorphic but still only accepts certain types), and
also get the type itself passed with the data.