Changed to JavaScriptInvoke to enable Property Nodes (#517)
Changes to the JavaScriptInvoke node to enable usage with the JavaScriptInvoke-based Property Node implementation.

Specifically this changes the behavior of internal JavaScriptInvoke calls to receive valueRefs instead of performing conversions inside the JavaScriptInvoke call. This allows internal functions to use the full vireo api to read and write values.

In addition the public facing JavaScriptInvoke conversions were replaced with a JSLI-specific type visitors. Long term this will make supporting more types for the JSLI node easier.