Fix strtol usage for ControlReference string
The second parameter of strtol is an address to write the end pointer of an integer instead of the expected length of the string. Incorrect usage seemed to be causing corruption of the result integer value.

Fixed the behavior by stack allocating a copy of the tag as a CString before calling strtol because strtol expects CString semantics.