Support Promises for JavaScriptInvoke calls (#530)
Allows users to return a promise that resolves to a value or an error instead of relying on the getCompletionCallback for async operations. Does not change any existing async behavior while adding the new Promise capability.

Also has a refactor to the async tests to ensure that the all the test cases cover both styles of async usage.