Fixing JS reference being written in a loop
Fixing the case of control reference wire being written in a loop. We should allow this case. Originally we thought we couldn't get to the scenario of changing the value of a JS reference, but we can easily get into this case if the wire was being re-written to a new value in a loop with Reference PropertyRead (ie, label reference).