Coerce returnValue to Error on error paths.
For both sync and promise execution we have different code paths used for reporting errors.
We now use that as a hint to coerce a non-error value that was thrown / rejected into an Error value to report correctly.
For completioncallback we don't have the distinction and the user is required to correctly handle the returnValue.