Fix errors in type comparison in variant nodes involving arrays and clusters (#591)
Report unsupported error for cluster wired to VariantToData node with array as type

A special case that we do not support currently is when an cluster is wired in to the VariantToData node expecting an array. Though unsupported today, this is a valid use-case and should not report back the usual error 91. In this change, I am changing the error code we return in this case to 1558, which is the closest existing error code we could find.