Add support for variant equality comparison (#609)
# DETAILS
Unlike current equality checks in Vireo, Variant equality cannot be decomposed into equality checks on underlying types during parse time. The decision should instead be made at run time.

# IMPLEMENTATION
This change introduces a type visitor that traverses through nested types and performs an operation on them to return a resultant value. In this case, the operation is an equality operation on the types. To make this infrastructure useful for type conversion work that is coming up, the visitor works in two passes. In the first pass, the types are traversed to make sure that they're compatible types. In the second pass, the data in these types are compared. (for conversion, the data values will be converted and copied over in the second pass instead)

# TESTING
I also added elaborate tests to verify that equality check works across scalar and container data types. For details on what the tests are doing, looking at the `TestResult_` log line in the vtr files will explain it. There are too many tests for me to screenshot here. If you would like to look at the gvi/gviweb files used to generate them, I can send them over.