Vireo data item needs update implementation (#645)
Adds a needsUpdate flag to all parameters and the locals marked as data items in Vireo. With this change the HTML Panel is able to ask Vireo if a control associated with a variable needs an update and can avoid reading the value if there is no change.

The outcome of this change is controls with large values that are not updated frequently can avoid unnecessary work if there are no writes to their data item.