Update dev dependencies to latest (#669)
Updates Vireo's devDependencies to latest to address eslint dependency vulnerability warnings.

Part of the update migrated to rollup 2 which requires node >= 10. As part of this change I updated the node versions used on the CIs and described in the README to the current node LTS 12.

The change does not update karma-jasmine to the latest due to mnasyrov/karma-jasmine-spec-tags#7 but other dependencies were brought up to latest.

In addition, minor changes were made to accommodate new eslint rules. Only test files required changes, no source files.