Remove Write to system log API from Vireo (#694)
The Write to system log API for the Vireo target is now built using JSLI in the NXG code base(it was previously using an SLI). 
So we no longer need this function definition in the Vireo code base. 

Note: 
I have also removed the tests related to the same function.
PR for changes on the NXG codebase: [PR](https://ni.visualstudio.com/DevCentral/_git/ASW/pullrequest/198533)