Bug fix for timeout event with 0ms doesn't work with webvi. (#698)
WebVI does not execute when an event structure is used with a timeout event of 0 ms.
BUG [link](https://ni.visualstudio.com/DevCentral/_workitems/edit/1474951)
Whenever event structure timeout was 0ms it goes to InitObservableTimerState with tickCount as 0x7FFFFFFFFFFFFFFF setting it to infinite wait instead of executing event structure immediately.
![1](https://user-images.githubusercontent.com/97939637/152923175-6637ef0a-9691-46e1-b215-3470675e1b2c.PNG)
At first, I tried adding a condition in ReserveObservationStatesWithTimeout to check if tickCount>= 0 , pass tickCount in InitObservableTimerState so that we immediately trigger the observer whenever the timeout is 0. But during testing we found that it affected the behavior of other nodes as well such as the OneButtonDialog.
So, didn't move forward with this approach.
Now, I have added a condition if the timeout given is 0ms we execute the event structure immediately by going to the desired branch target. I have tested it with different configurations and the behavior was the same as the current vireo version.

Have added test cases and have also tested with other test cases.