Expose private http client manager (#699)
# Description

This PR makes a trivial change of making it possible to access the internal HttpClientManger instance associated with Vireo. It does not make it part of the public API for Vireo but instead makes it possible to access via Vireo's private API.

The intention is to make it feasible to access the state of an HttpClient instance (the configured CORS state, username, password, headers) so that libraries written with the JSLI could access that information (at their own risk) to extend the HttpClient with new nodes (Post Multipart, Patch, Get - Binary, etc).

In addition, this PR pins the httpbin dependencies so that the httpbin server can be installed reliably for testing.

## Testing

The PR includes a test showing how the internal private HttpClientManager and HttpClient can be accessed and validates that the configured states of the HttpClient instance can be observed from the JS context given access to the vireo instance.

I did not run the full ASW test suite for this PR because the change has no functional changes, it only exposes a new property on the vireo private api.